Refactored backend

+++ src/main/java/com/montran/banking/security/JwtTokenUtil.java
@@ -17,7 +17,7 @@ import io.jsonwebtoken.SignatureAlgorithm;
 @Component
 public class JwtTokenUtil implements Serializable {
 	private static final long serialVersionUID = -2550185165626007488L;
-	public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60; // 5 hours
+	public static final long JWT_TOKEN_VALIDITY = (long) 5 * 60 * 60; // 5 hours
 	@Value("${jwt.secret}")
 	private String secret;
